feat(svc-api-stack): refactor cdk folder and services folder
