+++ lib/stacks/svc-api-gateway-stack/cdk/api/interfaces/shared/index.ts
