+++ lib/stacks/svc-api-gateway-stack/cdk/api/lambda-integrations/management/general-lambda-Integration.ts
