+++ lib/stacks/svc-api-gateway-stack/cdk/api/resources/index.ts
