+++ lib/stacks/svc-api-gateway-stack/cdk/api/resources/management/resources.ts
