+++ lib/stacks/svc-api-gateway-stack/cdk/dynamo/tables.ts
