+++ lib/stacks/svc-api-gateway-stack/cdk/dynamo/types.ts
