+++ lib/stacks/svc-api-gateway-stack/services/lambda-api-integrations/management/get-dependencies.ts
@@ -1,7 +1,8 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import { QueryStringParametersPagination } from "../../api/interfaces/shared";
-import { extractDataFromEvent } from "../../api/utils/pre-process-event";
-import { InternalErrorResponse500, SuccessResponse200 } from "../../api/utils/api-response";
+import { QueryStringParametersPagination } from "../../../cdk/api/interfaces/shared";
+import { extractDataFromEvent } from "../../utils/pre-process-event";
+import { InternalErrorResponse500, SuccessResponse200 } from "../../utils/api-response";
+
 
 
 interface QueryParamsExpected extends QueryStringParametersPagination {
@@ -17,12 +18,12 @@ export const handler = async (
       propertyToExtract: "queryStringParameters",
     });
     if (!params) {
-      throw new Error("No query string parameters");
+      throw new Error("No query string parameters into request");
     }
     const users = [
       {
-        user_id: 1111,
-        name: "Rodrigo García",
+        dependency_id: 1111,
+        name: "Tren ligero",
       },
     ];
     return SuccessResponse200({ data: users });
