+++ lib/stacks/svc-api-gateway-stack/services/utils/pre-process-event.ts
